+++ src/main/java/lotto/domain/lottery/InvalidLottoNumberException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.lottery;
 
 public class InvalidLottoNumberException extends RuntimeException {
     public InvalidLottoNumberException(String message) {
