+++ src/main/java/lotto/domain/lottery/InvalidWinningLottoException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.lottery;
 
 public class InvalidWinningLottoException extends RuntimeException {
     public InvalidWinningLottoException(final String message) {
