+++ src/main/java/lotto/domain/lottery/LottoNumber.java
@@ -4,8 +4,6 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
-import lotto.exception.InvalidLottoNumberException;
-
 public class LottoNumber implements Comparable<LottoNumber> {
     private static final int MIN_NUMBER = 1;
     private static final int MAX_NUMBER = 45;
