+++ src/main/java/lotto/domain/purchase/InvalidMoneyAmountException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.purchase;
 
 public class InvalidMoneyAmountException extends RuntimeException {
     public InvalidMoneyAmountException(String message) {
