+++ src/main/java/lotto/domain/purchase/InvalidPurchaseAmountException.java
@@ -1,4 +1,4 @@
-package lotto.exception;
+package lotto.domain.purchase;
 
 public class InvalidPurchaseAmountException extends RuntimeException {
     public InvalidPurchaseAmountException(String message) {
