+++ src/test/java/lotto/domain/lottery/WinningLottoTest.java
@@ -13,8 +13,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import lotto.exception.InvalidWinningLottoException;
-
 class WinningLottoTest {
     private LottoTicket lottoTicket;
